backend/requirements.txt
flask
flask-socketio
flask-cors
google-generativeai
pymongo[srv]
python-dotenv
cryptography
eventlet
